+++ TinkoffASDKUI/TinkoffASDKUI/Modules/AddNewCard/AddNewCardPresenter.swift
@@ -36,11 +36,6 @@ final class AddNewCardPresenter {
 extension AddNewCardPresenter: IAddNewCardPresenter {
 
     func viewAddCardTapped() {
-        let validationResult = cardfieldFactoryResult?.presenter.validateWholeForm()
-        guard validationResult?.cardNumberIsValid == true,
-              validationResult?.expirationIsValid == true,
-              validationResult?.cvcIsValid == true
-        else { return }
         addCard()
     }
 
@@ -88,20 +83,18 @@ extension AddNewCardPresenter {
             number: cardfieldPresenter.cardNumber,
             expiration: cardfieldPresenter.expiration,
             cvc: cardfieldPresenter.cvc,
-            resultCompletion: { result in
-                DispatchQueue.performOnMain { [weak self] in
-                    guard let self = self
-                    else { return }
+            resultCompletion: { [weak self] result in
+                guard let self = self
+                else { return }
 
-                    self.view?.hideLoadingState()
+                self.view?.hideLoadingState()
 
-                    switch result {
-                    case let .success(card):
-                        self.view?.closeScreen()
-                        self.view?.notifyAdded(card: card)
-                    case let .failure(error):
-                        self.handleAddCard(error: error)
-                    }
+                switch result {
+                case let .success(card):
+                    self.view?.closeScreen()
+                    self.view?.notifyAdded(card: card)
+                case let .failure(error):
+                    self.handleAddCard(error: error)
                 }
             }
         )