+++ TinkoffASDKUI/TinkoffASDKUITests/Infrastructure/Utils.swift
@@ -1,0 +1,12 @@
+//
+//  Utils.swift
+//  TinkoffASDKUI-Unit-Tests
+//
+//  Created by Ivan Glushko on 28.12.2022.
+//
+
+import Foundation
+
+extension Double {
+    static let testTimeout: Double = 1
+}